[back] tests: re-add a tests about test_language_detection
